third_party/blink/web_tests/platform/generic/printing/page-break-always-expected.txt

Test for page-break-before:always and page-break-after:always

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "firstPage" is 0
PASS: page number of "secondPage1" is 1
PASS: page number of "secondPage2" is 1
PASS: page number of "thirdPage" is 2
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
